feat(module6-task2): Добавил сортировку
